+++ src/joedb/io/joedb_push.cpp
@@ -55,8 +55,6 @@ namespace joedb
 
    while (true)
    {
-    journal.pull();
-
     if (journal.get_checkpoint_position() > server_checkpoint)
      server_checkpoint = connection->push(journal, server_checkpoint, false);
 
@@ -64,6 +62,7 @@ namespace joedb
      break;
 
     std::this_thread::sleep_for(std::chrono::seconds(1));
+    journal.pull();
    }
   }